Get the possible grouping variables from an experiment

Given x- and y-variables to plot, get a list of the possible variables to use in grouping the data for plotting. These variables are the independent variables other than the y-variable.

get_group_vars(experiment, x_var, y_var)

Arguments

experiment

An experiment object, returned by load_bs_file(), load_bs_table(), or load_bs_spreadsheet().

x_var

The name of the x-variable to use in the plot.

y_var

The name of the y-variable to use in the plot.

Value

A data frame with the names of the possible grouping variables and the corresponding columns in the experiment data frame.